Update download links for 2.1.0 and beyond, and add deb pkg link.
We add the dedicated deb link because there is always going to be a
chicken and egg problem with Mender itself and the deb package: We
need to build a new demo artifact to verify the package, using the new
Mender version (for the artifact generators), but the deb package will
only be uploaded later. In the cases where they differ, this will lead
to two different test scenarios, one with the old one on Raspbian, and
one with the new one on QEMU/Yocto.

We also update to a temporary build tag, but we will switch to the
real one shortly.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>