Make sure that errors are caught while starting the service.
Since we use a script, systemd considers the service successfully
launched as soon as this script runs. However, we want to delay
readiness until we know that the busybox server actually runs,
otherwise we get the error only later, when it is too late to act on
it.

Changelog: Title

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>